refactor(parser): split builtin signatures registry
